Consolidate Addresspoints Status in client project
