+++ addresspoints/src/main/scala/grasshopper/addresspoints/api/Service.scala
@@ -18,14 +18,14 @@ import feature.Feature
 import grasshopper.addresspoints.search.Geocode
 import grasshopper.model.Status
 import grasshopper.model.addresspoints.{ AddressPointsResult, AddressInput }
-import grasshopper.protocol.addresspoints.AddressPointJsonProtocol
+import grasshopper.protocol.addresspoints.AddressPointsJsonProtocol
 import org.elasticsearch.client.Client
 import org.slf4j.LoggerFactory
 import spray.json._
 
 import scala.concurrent.ExecutionContextExecutor
 
-trait Service extends AddressPointJsonProtocol with Geocode {
+trait Service extends AddressPointsJsonProtocol with Geocode {
   implicit val system: ActorSystem
 
   implicit def executor: ExecutionContextExecutor
