+++ geocoder/src/main/scala/grasshopper/geocoder/model/GeocodeStatus.scala
@@ -1,7 +1,6 @@
 package grasshopper.geocoder.model
 
-import grasshopper.client.addresspoints.model.AddressPointsStatus
 import grasshopper.model.Status
 import grasshopper.client.parser.model.ParserStatus
 
-case class GeocodeStatus(addressPointsStatus: AddressPointsStatus, censusStatus: Status, parserStatus: ParserStatus)
+case class GeocodeStatus(addressPointsStatus: Status, censusStatus: Status, parserStatus: ParserStatus)
