+++ model/src/main/scala/grasshopper/protocol/addresspoints/AddressPointsJsonProtocol.scala
@@ -5,7 +5,7 @@ import grasshopper.model.addresspoints.{ AddressInput, AddressPointsResult }
 import spray.json.DefaultJsonProtocol
 import io.geojson.FeatureJsonProtocol._
 
-trait AddressPointJsonProtocol extends DefaultJsonProtocol {
+trait AddressPointsJsonProtocol extends DefaultJsonProtocol {
   implicit val statusFormat = jsonFormat4(Status.apply)
   implicit val addressInputFormat = jsonFormat1(AddressInput.apply)
   implicit val addressPointResult = jsonFormat3(AddressPointsResult.apply)
