+++ model/src/test/scala/grasshopper/protocol/addresspoints/AddressPointJsonProtocolSpec.scala
@@ -7,11 +7,11 @@ import grasshopper.model.Status
 import grasshopper.model.addresspoints.AddressInput
 import spray.json._
 
-class AddressPointJsonProtocolSpec extends FlatSpec with MustMatchers with AddressPointJsonProtocol {
+class AddressPointJsonProtocolSpec extends FlatSpec with MustMatchers with AddressPointsJsonProtocol {
 
   "Status" should "convert to and from JSON" in {
     val date = Calendar.getInstance().getTime().toString
-    val status = Status("OK", "grasshopper-grasshopper.addresspoints", date, InetAddress.getLocalHost.getHostName)
+    val status = Status("OK", "grasshopper-addresspoints", date, InetAddress.getLocalHost.getHostName)
     status.toJson.convertTo[Status] mustBe status
   }
